% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_workflow_adjust_basepop.R
\name{census_workflow_adjust_basepop}
\alias{census_workflow_adjust_basepop}
\title{census_workflow_adjust_basepop}
\usage{
census_workflow_adjust_basepop(
  popM1,
  popF1,
  popM_unsmoothed,
  popF_unsmoothed,
  Age_unsmoothed,
  smooth_method = NA,
  LocID,
  census_reference_date,
  nLxMatFemale = NULL,
  nLxMatMale = NULL,
  nLxMatDatesIn = NULL,
  AsfrMat = NULL,
  AsfrDatesIn = NULL,
  SRB = NULL,
  SRBDatesIn = NULL,
  radix = NULL
)
}
\arguments{
\item{popM1}{numeric. vector of male population counts by single year of age}

\item{popF1}{numeric. vector of female population counts by single year of age}
}
\value{
A list of objects describing the steps implemented in adjusting for missing children
}
\description{
This function carries out the basepop adjustment for children missing from census counts
as described in protocol flowchart xx
}
