% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_workflow_adjust_pes.R
\name{census_workflow_adjust_pes}
\alias{census_workflow_adjust_pes}
\title{census_workflow_adjust_pes}
\usage{
census_workflow_adjust_pes(
  Age,
  popF,
  popM,
  census_reference_date,
  NCE_total = NULL,
  NCE_m = NULL,
  NCE_f = NULL,
  NCE_age = NULL
)
}
\arguments{
\item{popF}{numeric. vector of female population counts}

\item{popM}{numeric. vector of male population counts}
}
\value{
A data frame with adjusted population counts by age
}
\description{
Adjust census populations per post-enumeration survey or user-defined adjustment factors
as described in protocol flowchart xx
}
