% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_workflow_adjust_smooth.R
\name{census_workflow_adjust_smooth}
\alias{census_workflow_adjust_smooth}
\title{census_workflow_adjust_smooth}
\usage{
census_workflow_adjust_smooth(
  popM,
  popF,
  Age,
  bachi_age_child = 3:17,
  bachi_age_adult = 23:77,
  age_ratio_age_child = c(0, 10),
  age_ratio_age_adult = c(15, 70),
  EduYrs
)
}
\arguments{
\item{popM}{numeric. vector of male population counts}

\item{popF}{numeric. vector of female population counts}

\item{Age}{numeric. vector of age (years) associated with population counts vectors}
}
\value{
A list of objects describing the steps implemented in smoothing population counts over age
}
\description{
this function assesses age heaping and accordingly applies smoothing
as described in protocol flowchart xx
}
