% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_workflow_for_one_census.R
\name{census_workflow_for_one_census}
\alias{census_workflow_for_one_census}
\title{census_workflow_for_one_census}
\usage{
census_workflow_for_one_census(
  dd_census_extract,
  LocID,
  locid_PES = LocID,
  locid_DemoTools = LocID,
  census_reference_date,
  adjust_pes = TRUE,
  adjust_smooth = TRUE,
  adjust_basepop = TRUE,
  lxMale = NULL,
  lxFemale = NULL,
  Age_lx = NULL,
  nLxMatMale = NULL,
  nLxMatFemale = NULL,
  nLxMatDatesIn = NULL,
  AsfrMat = NULL,
  AsfrDatesIn = NULL,
  SRB = NULL,
  SRBDatesIn = NULL,
  radix = NULL,
  EduYrs = NULL,
  PES_factor_Total = NULL,
  PES_factor_single_Male = NULL,
  PES_factor_single_Female = NULL,
  PES_factor_abridged_Male = NULL,
  PES_factor_abridged_Female = NULL
)
}
\arguments{
\item{dd_census_extract}{data frame output by ddharmony::DDharmonize_validate_PopCounts() function for one census}
}
\value{
A list of objects describing the steps implemented in the census adjustment protocol
}
\description{
implements the census adjustment protocol workflow for a single census
as described in protocol flowchart xx
}
\examples{
\dontrun{
LocID <- 312
load("C:/Users/SARAH/OneDrive - United Nations/WPP2021/Population/data/harmonized/312_census_harmonized.rda")
census_all <- pop \%>\% dplyr::filter(non_standard == FALSE & AgeStart >= 0)

dcids <- unique(census_all$DataCatalogID)
census_pop <- census_all \%>\% dplyr::filter(DataCatalogID == dcids[2])

census_reference_date <- as.numeric(census_pop$TimeMid[1])
census_reference_year <- ifelse(census_reference_date >= 1950, floor(census_reference_date), 1950)

dd_census_extract <- census_pop


test <- census_workflow_for_one_census (dd_census_extract, # a census extract returned by DDharmonize_validate_PopCounts()
                                        LocID,
                                        locid_DemoTools = LocID,
                                        census_reference_date, # decimal year
                                        adjust_pes = TRUE, # should census be adjusted based on pes models?
                                        adjust_smooth = TRUE, # should census be smoothed according to age heaping assessment?
                                        adjust_basepop = TRUE, # should child counts be adjusted per basepop analysis?
                                        lxMale = NULL, # single or abridged male lx at census reference year. if NULL then will use DemoToolsData
                                        lxFemale = NULL, # single or abridged female lx at census reference year.
                                        Age_lx = NULL, # single or abridged
                                        nLxMatMale = NULL, # matrix of abridged nLx for males. if NULL then will use DemoToolsData
                                        nLxMatFemale = NULL, # abridged
                                        nLxMatDatesIn = NULL,
                                        AsfrMat = NULL, # 5-year age groups from 15 to 45 only
                                        AsfrDatesIn = NULL,
                                        SRB = NULL,
                                        SRBDatesIn = NULL,
                                        radix = NULL,
                                        EduYrs = NULL,
                                        PES_factor_Total = NULL,
                                        PES_factor_single_Male = NULL,
                                        PES_factor_single_Female = NULL,
                                        PES_factor_abridged_Male = NULL,
                                        PES_factor_abridged_Female = NULL)
}
}
